update image_path to image
