+++ Backend/helpers/index.js
@@ -2,3 +2,4 @@ export * from './getProductApi.js'
 export * from './getProductImage.js'
 export * from './getProductTitle.js'
 export * from './removeProductImage.js'
+export * from './uploadProductImage.js'
